LOD, model reorganisation
git-svn-id: https://pioneer.svn.sourceforge.net/svnroot/pioneer/trunk@24 e632f14b-6550-0410-b89e-a82653faca30

+++ src/sbre/sbre.h
@@ -37,7 +37,7 @@ struct ObjParams
 };
 
 
-void sbreSetViewport (int w, int h, int d, float zn, float zf, float dn, float df);
+void sbreSetViewport (int w, int h, float d, float zn, float zf, float dn, float df);
 void sbreSetDirLight (float *pColor, float *pDir);
 void sbreSetWireframe (int val);
 void sbreRenderModel (Vector *pPos, Matrix *pOrient, int model, ObjParams *pParam,
